+++ src/modules/products/components/widgets/index.ts
@@ -1,2 +1,3 @@
 export { OffersWidget } from "./offers";
 export { AssetsWidget } from "./assets";
+export { VideosWidget } from "./videos";