+++ src/modules/products/components/widgets/videos/videos-widget.vue
@@ -3,17 +3,16 @@
     v-loading="loading"
     v-bind="props"
     :value="count"
-    :title="$t('PRODUCTS.PAGES.DETAILS.WIDGETS.OFFERS')"
-    icon="fas fa-tags"
+    :title="$t('PRODUCTS.PAGES.DETAILS.WIDGETS.VIDEOS')"
+    icon="fas fa-film"
     @click="clickHandler"
-  >
-  </VcWidget>
+  ></VcWidget>
 </template>
 
 <script setup lang="ts">
 import { VcWidget, useApiClient, useAsync, useBladeNavigation } from "@vc-shell/framework";
 import { UnwrapNestedRefs, onMounted, ref } from "vue";
-import { VcmpSellerCatalogClient, ISearchOffersQuery, SearchOffersQuery } from "vc-vendor-portal-api/marketplacevendor";
+import { CatalogModuleVideosClient, IVideoSearchCriteria, VideoSearchCriteria } from "vc-vendor-portal-api/catalog";
 
 interface Props {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -22,7 +21,7 @@ interface Props {
 
 const props = defineProps<Props>();
 
-const { getApiClient } = useApiClient(VcmpSellerCatalogClient);
+const { getApiClient } = useApiClient(CatalogModuleVideosClient);
 const client = getApiClient();
 
 const { openBlade, resolveBladeByName } = useBladeNavigation();
@@ -33,10 +32,8 @@ const count = ref(0);
 function clickHandler() {
   if (!widgetOpened.value) {
     openBlade({
-      blade: resolveBladeByName("Offers"),
-      options: {
-        sellerProduct: props.modelValue?.item,
-      },
+      blade: resolveBladeByName("VideosList"),
+      param: props.modelValue?.item?.publishedProductDataId ?? props.modelValue?.item?.stagedProductDataId,
       onOpen() {
         widgetOpened.value = true;
       },
@@ -47,15 +44,17 @@ function clickHandler() {
   }
 }
 
-const { loading, action: getCount } = useAsync<ISearchOffersQuery, number | undefined>(async (query) => {
-  return (await client).searchOffers(new SearchOffersQuery(query)).then((res) => res.totalCount);
+const { loading, action: getCount } = useAsync<IVideoSearchCriteria, number | undefined>(async (query) => {
+  return (await client).searchVideos(new VideoSearchCriteria(query)).then((res) => res.totalCount);
 });
 
 onMounted(async () => {
   if (props.modelValue?.item?.id) {
     count.value = await getCount({
       take: 0,
-      sellerProductId: props.modelValue?.item?.id,
+      ownerIds: props.modelValue?.item?.publishedProductDataId
+        ? [props.modelValue?.item?.publishedProductDataId]
+        : [props.modelValue?.item?.stagedProductDataId],
     });
   }
 });